+++ src/lib/client/index.ts
@@ -35,6 +35,7 @@ export {
 	setError,
 	withFiles,
 	removeFiles,
+	fail,
 	type SuperValidated,
 	type TaintedFields,
 	type ValidationErrors