+++ src/routes/(v2)/v2/Navigation.svelte
@@ -55,7 +55,8 @@
 		'files-proxy',
 		'tainted-array',
 		'spa-clearonsubmit',
-		'app-error'
+		'app-error',
+		'form-result-type'
 	].sort();
 </script>