+++ src/routes/(v2)/v2/form-result-type/schema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+	dir: z
+		.enum(['', 'north', 'south', 'east', 'west'])
+		.refine((dir) => dir != '', 'No direction specified')
+});