+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- The `ActionResult` for success or failure is now added to the `onUpdate` event in the `result` property.
+- Added a `fail` function, works the same as the SvelteKit fail, but removes files and sets `form.valid` to `false`.
+
 ## [2.10.6] - 2024-03-20
 
 ### Changed